Refactor application structure by moving app configuration and main component files to appropriate directories
