+++ apps/web/questlog-web/src/app/presentation/app.ts
@@ -1,12 +1,12 @@
 import { Component, signal, inject, OnInit } from '@angular/core';
 import { RouterModule, RouterOutlet } from '@angular/router';
-import { AppStore } from './application/app.store';
+import { AppStore } from '../application/app.store';
 
 @Component({
   selector: 'app-root',
   imports: [RouterModule, RouterOutlet],
-  templateUrl: './presentation/app.html',
-  styleUrl: './presentation/app.scss'
+  templateUrl: 'app.html',
+  styleUrl: 'app.scss'
 })
 
 export class App implements OnInit{
