+++ apps/web/questlog-web/src/main.ts
@@ -1,6 +1,6 @@
 import { bootstrapApplication } from '@angular/platform-browser';
-import { appConfig } from './app/presentation/app.config';
-import { App } from './app/app';
+import { appConfig } from './app/app.config';
+import { App } from './app/presentation/app';
 import 'bootstrap/dist/js/bootstrap.bundle.min.js';
 
 bootstrapApplication(App, appConfig)
